updating calendar page to most recent and reverting settings
